Update logger.py
`import phonemizer` cause a call to `get_logger()` which configure the default logger. This cause duplicated logs if the same logger is used in the project importing Phonemizer. To avoid this, we have to call `get_logger('quiet')` after importing. It might be easier to name the logger used in Phonemizer, and to have it quiet by default.